refactor(oracle): migrate wifi dot11 with phase-15 stack encoder
