+++ tools/oracle/engine/generator.py
@@ -23,7 +23,6 @@ from .sampling import (
     _SkipField,
     _string_or_none,
     _declared_ethertype_for_stack,
-    _dot11_is_management,
     _different_ipv4,
     _different_ipv6,
     _different_mac,
@@ -35,7 +34,6 @@ from .sampling import (
     _IPV6_DOCUMENTATION_NETWORK,
     _ipv6_next_header_for_stack,
     _json_object,
-    _mac_for_domain,
     _next_layer_after,
     _object,
     _object_list,
@@ -106,12 +104,10 @@ from .protocols.dns import _dns_behavior_emits_raw
 # the cross-backend option matrix — the same co-locate-and-re-import pattern as
 # the DNS ``_dns_behavior_emits_raw`` helper above.
 from .protocols.dhcp import DHCP_OPTION_MATRIX_TOKENS
-# ``_rsn_information_value_hex`` moved to the Wi-Fi sampler plugin
-# (``protocols/wifi.py``) with the ``eapol`` and ``rsn`` samplers. It is
-# re-imported here because the still-resident ``dot11`` management-tag sampler
-# (``_dot11_management_tags``) builds the RSN information element from it; the
-# helper moves out of the generator entirely once ``dot11`` migrates.
-from .protocols.wifi import _rsn_information_value_hex
+# The ``radiotap`` / ``dot11`` / ``eapol`` / ``rsn`` samplers and the
+# ``_rsn_information_value_hex`` helper all live in the Wi-Fi sampler plugin
+# (``protocols/wifi.py``); the generator reaches them through ``SAMPLER_REGISTRY``
+# and no longer re-imports any 802.11 helper.
 
 
 SUPPORTED_LAYER_BACKEND = "libcrafter"
@@ -119,34 +115,10 @@ SUPPORTED_LAYER_BACKEND = "libcrafter"
 _DERIVED_DOMAINS = {"derived"}
 
 _SUPPORTED_FIELDS: dict[str, set[str]] = {
-    "dot11": {
-        "addr1",
-        "addr2",
-        "addr3",
-        "addr4",
-        "duration_id",
-        "frame_control",
-        "frame_type",
-        "from_ds",
-        "ht_control",
-        "management_fixed_fields",
-        "more_data",
-        "more_fragments",
-        "order",
-        "payload",
-        "power_management",
-        "protected",
-        "protocol_version",
-        "qos_control",
-        "retry",
-        "sequence_control",
-        "subtype",
-        "tagged_parameters",
-        "to_ds",
-    },
-    # The ``eapol`` and ``rsn`` field allowlists moved to ``protocols/wifi.py``
-    # (their ``ProtocolSampler.supported_fields``); ``_supported_fields`` resolves
-    # them from the registry. ``dot11`` stays here until it migrates next.
+    # The ``dot11`` / ``eapol`` / ``radiotap`` / ``rsn`` field allowlists moved to
+    # ``protocols/wifi.py`` (their ``ProtocolSampler.supported_fields``);
+    # ``_supported_fields`` resolves them from the registry. Every Wi-Fi layer is
+    # now registered, so this table no longer carries any 802.11 entry.
 }
 
 
@@ -1662,8 +1634,6 @@ class PacketGenerator:
                 field_spec=field_spec,
                 current_fields=current_fields,
             )
-        if layer == "dot11":
-            return _sample_dot11_field(ctx, field_name, domain, current_fields)
 
         raise ValueError(f"spec error: unsupported layer sampler: {layer}")
 
@@ -2255,196 +2225,11 @@ def _auto_sample_feature(feature: str) -> bool:
     return feature != "icmpv4_errors"
 
 
-def _sample_dot11_field(
-    ctx: _SamplingContext,
-    field_name: str,
-    domain: object,
-    current_fields: Mapping[str, object],
-) -> object:
-    frame_control = _dot11_frame_control_for_case(ctx.case, ctx.stack)
-    if field_name == "frame_control":
-        return frame_control
-    if field_name == "protocol_version":
-        return frame_control & 0x3
-    if field_name == "frame_type":
-        return (frame_control >> 2) & 0x3
-    if field_name == "subtype":
-        return (frame_control >> 4) & 0xF
-    if field_name == "to_ds":
-        return 1 if frame_control & 0x0100 else 0
-    if field_name == "from_ds":
-        return 1 if frame_control & 0x0200 else 0
-    if field_name == "more_fragments":
-        return 0
-    if field_name == "retry":
-        return 0
-    if field_name == "power_management":
-        return 0
-    if field_name == "more_data":
-        return 0
-    if field_name == "protected":
-        return 1 if frame_control & 0x4000 else 0
-    if field_name == "order":
-        return 0
-    if field_name == "duration_id":
-        if domain == "nav":
-            return 314
-        if domain == "association_id":
-            return 0xC001
-        if domain == "boundary":
-            return 0xFFFF
-        return 0
-    if field_name == "addr1":
-        return _mac_for_domain(ctx, domain, ctx.dst_mac)
-    if field_name == "addr2":
-        if domain == "absent" and _dot11_control_address_count(frame_control) < 2:
-            return _SKIP_FIELD
-        return _mac_for_domain(ctx, domain, ctx.src_mac)
-    if field_name == "addr3":
-        if domain == "absent" and _dot11_header_has_three_addresses(frame_control):
-            return ctx.dst_mac
-        if domain == "absent":
-            return _SKIP_FIELD
-        return _mac_for_domain(ctx, domain, "00:00:5e:00:53:03")
-    if field_name == "addr4":
-        if not _dot11_has_addr4(current_fields):
-            return _SKIP_FIELD
-        return _mac_for_domain(ctx, domain, "00:00:5e:00:53:04")
-    if field_name == "sequence_control":
-        if not _dot11_header_has_sequence_control(frame_control):
-            return _SKIP_FIELD
-        if _dot11_is_data(frame_control) and "llc_snap" in ctx.stack:
-            if domain == "sequence":
-                return 0x1230
-            if domain == "boundary":
-                return 0xFFF0
-            return 0x1000
-        if domain == "absent":
-            return 0x1000
-        if domain == "fragment_zero":
-            return 0x1000
-        if domain == "sequence":
-            return 0x1230
-        if domain == "boundary":
-            return 0xFFFF
-        return 0x1000
-    if field_name == "qos_control":
-        if not _dot11_has_qos_control(frame_control):
-            return _SKIP_FIELD
-        if domain == "ack_policy":
-            return 0x0020
-        if domain == "boundary":
-            return 0xFFFF
-        return 0
-    if field_name == "ht_control":
-        return _SKIP_FIELD
-    if field_name == "management_fixed_fields":
-        if not _dot11_is_management(frame_control):
-            return _SKIP_FIELD
-        return {"hex": _dot11_management_fixed_hex(frame_control, str(domain))}
-    if field_name == "tagged_parameters":
-        subtype = (frame_control >> 4) & 0x0F
-        if not _dot11_is_management(frame_control) or not _dot11_management_subtype_has_tags(subtype):
-            return _SKIP_FIELD
-        return _dot11_management_tags(ctx.case, str(domain))
-    if field_name == "payload":
-        return _SKIP_FIELD
-    raise ValueError(f"spec error: unsupported dot11 field sampler: {field_name}")
-
-
-def _dot11_management_fixed_hex(frame_control: int, domain: str) -> str:
-    subtype = (frame_control >> 4) & 0x0F
-    if subtype in {5, 8}:  # probe response, beacon
-        return "000000000000000064000100"
-    if subtype == 0:  # association request
-        return "31040000"
-    if subtype in {1, 3}:  # association/reassociation response
-        return "310400000100"
-    if subtype == 2:  # reassociation request
-        return "3104000000005e005301"
-    if subtype == 11:  # authentication
-        return "000001000000"
-    if subtype in {10, 12}:  # disassociation/deauthentication
-        return "0100"
-    if subtype in {13, 14}:  # action/action-no-ack category
-        return "00"
-    if domain == "association_fixed":
-        return "31040000"
-    if domain == "authentication_fixed":
-        return "000001000000"
-    return ""
-
-
-def _dot11_management_tags(case: str, domain: str) -> list[JSONObject]:
-    if domain == "absent":
-        return []
-    if domain == "rsn" or "rsn" in case.replace("_", "-"):
-        return [{"id": 48, "value": {"hex": _rsn_information_value_hex()}}]
-    if domain == "supported_rates":
-        return [{"id": 1, "value": {"hex": "82848b96"}}]
-    if domain == "unknown":
-        return [{"id": 221, "value": {"hex": "0050f20101"}}]
-    return [
-        {"id": 0, "value": {"hex": "6c696263726166746572"}},
-        {"id": 1, "value": {"hex": "82848b96"}},
-        {"id": 221, "value": {"hex": "0050f20101"}},
-    ]
-
-
-def _dot11_management_subtype_has_tags(subtype: int) -> bool:
-    return subtype in {0, 2, 4, 5, 8}
-
-
-# The ``eapol`` and ``rsn`` field samplers and the ``_rsn_information_value_hex``
-# helper moved to ``protocols/wifi.py`` with the registered ``ProtocolSampler``s;
-# ``_rsn_information_value_hex`` is re-imported above for the still-resident
-# ``dot11`` management-tag sampler.
-
-
-def _dot11_frame_control_for_case(case: str, stack: Sequence[str]) -> int:
-    key = case.replace("_", "-")
-    if "llc_snap" in stack:
-        return 2 << 2
-    if "control" in key:
-        return (1 << 2) | (11 << 4)
-    if "management" in key or "rsn" in key:
-        return (0 << 2) | (8 << 4)
-    subtype = 8 if "qos" in key else 0
-    flags = 0x4000 if "protected" in key else 0
-    if "tods-fromds" in key:
-        flags |= 0x0300
-    return (2 << 2) | (subtype << 4) | flags
-
-
-def _dot11_is_control(frame_control: int) -> bool:
-    return ((frame_control >> 2) & 0x3) == 1
-
-
-def _dot11_is_data(frame_control: int) -> bool:
-    return ((frame_control >> 2) & 0x3) == 2
-
-
-def _dot11_header_has_three_addresses(frame_control: int) -> bool:
-    return not _dot11_is_control(frame_control)
-
-
-def _dot11_header_has_sequence_control(frame_control: int) -> bool:
-    return not _dot11_is_control(frame_control)
-
-
-def _dot11_has_qos_control(frame_control: int) -> bool:
-    return ((frame_control >> 2) & 0x3) == 2 and (((frame_control >> 4) & 0xF) & 0x8) != 0
-
-
-def _dot11_has_addr4(fields: Mapping[str, object]) -> bool:
-    return bool(fields.get("to_ds")) and bool(fields.get("from_ds"))
-
-
-def _dot11_control_address_count(frame_control: int) -> int:
-    if not _dot11_is_control(frame_control):
-        return 3
-    subtype = (frame_control >> 4) & 0xF
-    return 1 if subtype in {12, 13} else 2
+# The ``dot11`` field sampler and its frame-control/management-tag helpers moved
+# to ``protocols/wifi.py`` with the registered ``ProtocolSampler`` (the last 802.11
+# layer to migrate). The generator now reaches the ``dot11`` sampler through
+# ``SAMPLER_REGISTRY``; nothing in the generator references the moved helpers or the
+# ``_rsn_information_value_hex`` builder anymore.
 
 
 def generate_plans(
