+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -42,7 +42,6 @@ KNOWN_UNMIGRATED: frozenset[str] = frozenset(
     {
         "ble_adv",
         "ble_radio",
-        "dot11",
     }
 )
